Update v_ksm_data_mart.sql
Performed the inner join for efficentcy.

Removed "Students" as a record type because I found alumni from other schools that are enrolled at Kellogg.

So for example: Someone could have graduated Weinberg in 2005 and now enrolled in an MBA program as a student. However, their record status is alumnus in CATs.